Revert "Merge branch 'dev' of git://github.com/Baystation12/Baystation12 into Baystation12-dev"
This reverts commit 1a40d528ac3fad762f21c02cfdf0583fce46b95e.

+++ code/game/objects/effects/decals/Cleanable/humans.dm
@@ -32,30 +32,27 @@ var/global/list/image/splatter_cache=list()
 	if(invisibility != 100)
 		invisibility = 100
 		amount = 0
-		GLOB.processing_objects -= src
+		processing_objects -= src
 	..(ignore=1)
 
-/obj/effect/decal/cleanable/blood/hide()
-	return
-
 /obj/effect/decal/cleanable/blood/Destroy()
-	GLOB.processing_objects -= src
+	processing_objects -= src
 	return ..()
 
-/obj/effect/decal/cleanable/blood/Initialize()
-	. = ..()
+/obj/effect/decal/cleanable/blood/New()
+	..()
 	update_icon()
 	if(istype(src, /obj/effect/decal/cleanable/blood/gibs))
 		return
 	if(src.type == /obj/effect/decal/cleanable/blood)
-		if(isturf(src.loc))
+		if(src.loc && isturf(src.loc))
 			for(var/obj/effect/decal/cleanable/blood/B in src.loc)
 				if(B != src)
 					if (B.blood_DNA)
 						blood_DNA |= B.blood_DNA.Copy()
 					qdel(B)
 	drytime = world.time + DRYING_TIME * (amount+1)
-	GLOB.processing_objects += src
+	processing_objects += src
 
 /obj/effect/decal/cleanable/blood/process()
 	if(world.time > drytime)
@@ -117,7 +114,7 @@ var/global/list/image/splatter_cache=list()
 	desc = drydesc
 	color = adjust_brightness(color, -50)
 	amount = 0
-	GLOB.processing_objects -= src
+	processing_objects -= src
 
 /obj/effect/decal/cleanable/blood/attack_hand(mob/living/carbon/human/user)
 	..()
@@ -245,6 +242,5 @@ var/global/list/image/splatter_cache=list()
 	var/dry=0 // Keeps the lag down
 
 /obj/effect/decal/cleanable/mucus/New()
-	..()
 	spawn(DRYING_TIME * 2)
 		dry=1
