+++ code/game/supplyshuttle.dm
@@ -142,7 +142,10 @@ var/list/point_source_descriptions = list(
 	var/list/master_supply_list = list()
 	//shuttle movement
 	var/movetime = 1200
-	var/datum/shuttle/autodock/ferry/supply/shuttle
+	var/datum/shuttle/ferry/supply/shuttle
+
+	var/obj/machinery/computer/supply/primaryterminal //terminal hardcopy forms will be printed to.
+
 
 	New()
 		ordernum = rand(1,9000)
@@ -177,37 +180,40 @@ var/list/point_source_descriptions = list(
 
 	//Sellin
 	proc/sell()
+		var/area/area_shuttle = shuttle.get_location_area()
+		if(!area_shuttle)	return
+
 		var/phoron_count = 0
 		var/plat_count = 0
-		for(var/area/subarea in shuttle.shuttle_area)
-			for(var/atom/movable/MA in subarea)
-				if(MA.anchored)	continue
 
-				// Must be in a crate!
-				if(istype(MA,/obj/structure/closet/crate))
-					var/obj/structure/closet/crate/CR = MA
-					callHook("sell_crate", list(CR, subarea))
+		for(var/atom/movable/MA in area_shuttle)
+			if(MA.anchored)	continue
 
-					add_points_from_source(CR.points_per_crate, "crate")
-					var/find_slip = 1
+			// Must be in a crate!
+			if(istype(MA,/obj/structure/closet/crate))
+				var/obj/structure/closet/crate/CR = MA
+				callHook("sell_crate", list(CR, area_shuttle))
 
-					for(var/atom in CR)
-						// Sell manifests
-						var/atom/A = atom
-						if(find_slip && istype(A,/obj/item/weapon/paper/manifest))
-							var/obj/item/weapon/paper/manifest/slip = A
-							if(!slip.is_copy && slip.stamped && slip.stamped.len) //yes, the clown stamp will work. clown is the highest authority on the station, it makes sense
-								add_points_from_source(points_per_slip, "manifest")
-								find_slip = 0
-							continue
+				add_points_from_source(CR.points_per_crate, "crate")
+				var/find_slip = 1
 
-						// Sell phoron and platinum
-						if(istype(A, /obj/item/stack))
-							var/obj/item/stack/P = A
-							switch(P.get_material_name())
-								if("phoron") phoron_count += P.get_amount()
-								if("platinum") plat_count += P.get_amount()
-				qdel(MA)
+				for(var/atom in CR)
+					// Sell manifests
+					var/atom/A = atom
+					if(find_slip && istype(A,/obj/item/weapon/paper/manifest))
+						var/obj/item/weapon/paper/manifest/slip = A
+						if(!slip.is_copy && slip.stamped && slip.stamped.len) //yes, the clown stamp will work. clown is the highest authority on the station, it makes sense
+							add_points_from_source(points_per_slip, "manifest")
+							find_slip = 0
+						continue
+
+					// Sell phoron and platinum
+					if(istype(A, /obj/item/stack))
+						var/obj/item/stack/P = A
+						switch(P.get_material_name())
+							if("phoron") phoron_count += P.get_amount()
+							if("platinum") plat_count += P.get_amount()
+			qdel(MA)
 
 		if(phoron_count)
 			var/temp = phoron_count * points_per_phoron
@@ -220,18 +226,20 @@ var/list/point_source_descriptions = list(
 	//Buyin
 	proc/buy()
 		if(!shoppinglist.len) return
+		var/area/area_shuttle = shuttle.get_location_area()
+		if(!area_shuttle)	return
 		var/list/clear_turfs = list()
-		for(var/area/subarea in shuttle.shuttle_area)
-			for(var/turf/T in subarea)
-				if(T.density)	continue
-				var/contcount
-				for(var/atom/A in T.contents)
-					if(!A.simulated)
-						continue
-					contcount++
-				if(contcount)
+
+		for(var/turf/T in area_shuttle)
+			if(T.density)	continue
+			var/contcount
+			for(var/atom/A in T.contents)
+				if(!A.simulated)
 					continue
-				clear_turfs += T
+				contcount++
+			if(contcount)
+				continue
+			clear_turfs += T
 		for(var/S in shoppinglist)
 			if(!clear_turfs.len)	break
 			var/i = rand(1,clear_turfs.len)
@@ -252,7 +260,7 @@ var/list/point_source_descriptions = list(
 				slip.is_copy = 0
 				slip.info = "<h3>[command_name()] Shipping Manifest</h3><hr><br>"
 				slip.info +="Order #[SO.ordernum]<br>"
-				slip.info +="Destination: [GLOB.using_map.station_name]<br>"
+				slip.info +="Destination: [using_map.station_name]<br>"
 				slip.info +="[shoppinglist.len] PACKAGES IN THIS SHIPMENT<br>"
 				slip.info +="CONTENTS:<br><ul>"
